Extracts logic to start the app into it's own app.js file - Useful for upcoming testing as this seperates listenting port
